requirements.txt
# requirements.txt

# Flask and Web Server
flask
flask_cors
python-dotenv

# LangChain and RAG
langchain-google-genai
langchain-community
pypdf
chromadb

# AI and Machine Learning Core
sentence-transformers
scikit-learn
numpy<2  # <--- THIS IS THE CRITICAL FIX

# The pandas library is an indirect dependency, 
# but we don't need to list it here. Pip will handle it.